+++ GeosCore/wetscav_mod.F90
@@ -1697,9 +1697,10 @@ CONTAINS
     UNITCHANGE_KGKG = .FALSE.
     UNITCHANGE_KGM2 = .FALSE.
 
-    IF ( TRIM( State_Chm%Spc_Units ) .eq. 'kg/kg dry' ) THEN
+    IF ( State_Chm%Spc_Units == KG_SPECIES_PER_KG_DRY_AIR ) THEN
        UNITCHANGE_KGKG = .TRUE.
-       CALL ConvertBox_KgKgDry_to_Kg( I, J, L, State_Met, State_Chm, .FALSE., RC )
+       CALL ConvertBox_KgKgDry_to_Kg( I,         J,          L,              &
+                                      State_Met, State_Chm, .FALSE., RC     )
 
        ! Trap potential errors
        IF ( RC /= GC_SUCCESS ) THEN
@@ -1708,7 +1709,7 @@ CONTAINS
           RETURN
        ENDIF
 
-    ELSE IF ( TRIM( State_Chm%Spc_Units ) .eq. 'kg/m2' ) THEN
+    ELSE IF ( State_Chm%Spc_Units == KG_SPECIES_PER_M2 ) THEN
        UNITCHANGE_KGM2 = .TRUE.
        CALL ConvertBox_Kgm2_to_Kg( I, J, L, State_Chm, State_Grid, .FALSE., RC )
 
@@ -1723,7 +1724,7 @@ CONTAINS
 
        ! Exit if units are not as expected
        ErrMsg = 'Incorrect initial species units:' // &
-                TRIM( State_Chm%Spc_Units )
+                TRIM( UNIT_STR( State_Chm%Spc_Units ) )
        CALL GC_Error( ErrMsg, RC, ThisLoc )
        RETURN
 
@@ -1994,10 +1995,10 @@ CONTAINS
     ENDIF
 
     ! Check that species units are as expected (ewl, 9/29/15)
-    IF ( TRIM( State_Chm%Spc_Units ) /= 'kg/kg dry' .AND. &
-         TRIM( State_Chm%Spc_Units ) /= 'kg/m2' ) THEN
+    IF ( State_Chm%Spc_Units /= KG_SPECIES_PER_KG_DRY_AIR  .AND.              &
+         State_Chm%Spc_Units /= KG_SPECIES_PER_M2         ) THEN
        ErrMsg = 'Incorrect final species units:' // &
-                TRIM( State_Chm%Spc_Units )
+                TRIM( UNIT_STR(State_Chm%Spc_Units) )
        CALL GC_Error( ErrMsg, RC, ThisLoc )
        RETURN
     ENDIF
@@ -3041,7 +3042,7 @@ CONTAINS
     USE State_Met_Mod,    ONLY : MetState
     USE TIME_MOD,         ONLY : GET_TS_DYN
     USE Species_Mod,      ONLY : Species
-    USE UnitConv_Mod,     ONLY : Convert_Spc_Units
+    USE UnitConv_Mod
 !
 ! !INPUT PARAMETERS:
 !
@@ -3184,6 +3185,7 @@ CONTAINS
     LOGICAL                :: IS_RAINOUT, IS_WASHOUT, IS_BOTH
     INTEGER                :: I,     IDX,    J,         L
     INTEGER                :: N,     NW,     Hg_Cat,    EC
+    INTEGER                :: origUnit
     REAL(fp)               :: Q,     QDOWN,  DT,        DT_OVER_TAU
     REAL(fp)               :: K,     K_MIN,  K_RAIN,    RAINFRAC
     REAL(fp)               :: F,     FTOP,   F_PRIME,   WASHFRAC
@@ -3201,7 +3203,6 @@ CONTAINS
                                    State_Grid%NZ,State_Grid%NX,State_Grid%NY)
 
     ! Strings
-    CHARACTER(LEN=63)      :: OrigUnit
     CHARACTER(LEN=255)     :: ErrMsg, ErrorMsg, ThisLoc
 
     ! Pointers
@@ -3225,8 +3226,14 @@ CONTAINS
 
     ! Convert species concentration to mass per unit area (kg/m2) for
     ! wet deposition since computation is done per column (ewl, 9/8/15)
-    CALL Convert_Spc_Units( Input_Opt, State_Chm, State_Grid, State_Met, &
-                            'kg/m2', RC, OrigUnit=OrigUnit )
+    CALL Convert_Spc_Units(                                                  &
+         Input_Opt  = Input_Opt,                                             &
+         State_Chm  = State_Chm,                                             &
+         State_Grid = State_Grid,                                            &
+         State_Met  = State_Met,                                             &
+         outUnit    = KG_SPECIES_PER_M2,                                     &
+         origUnit   = origUnit,                                              &
+         RC         = RC                                                    )
 
     ! Trap potential errors
     IF ( RC /= GC_SUCCESS ) THEN
@@ -3803,8 +3810,13 @@ CONTAINS
     ENDIF
 
     ! Convert species concentration back to original unit (ewl, 9/8/15)
-    CALL Convert_Spc_Units( Input_Opt, State_Chm, State_Grid, State_Met, &
-                            OrigUnit,  RC )
+    CALL Convert_Spc_Units(                                                  &
+         Input_Opt  = Input_Opt,                                             &
+         State_Chm  = State_Chm,                                             &
+         State_Grid = State_Grid,                                            &
+         State_Met  = State_Met,                                             &
+         outUnit    = origUnit,                                              &
+         RC         = RC                                                    )
 
     ! Trap potential errors
     IF ( RC /= GC_SUCCESS ) THEN
@@ -4288,6 +4300,7 @@ CONTAINS
 #ifdef TOMAS
     USE TOMAS_MOD,      ONLY : IBINS, ICOMP, AQOXID
 #endif
+    USE UnitConv_Mod
 !
 ! !INPUT PARAMETERS:
 !
@@ -4609,13 +4622,13 @@ CONTAINS
                 ! therefore converted to [kg] locally within AQOXID.
                 ! GAINED is now [kg/m2] ans so is multiplied
                 ! by area prior to passing REEVAPSO2 to AQOXID (ewl, 9/30/15)
-                IF ( TRIM( State_Chm%Spc_Units ) .eq. 'kg/m2' ) THEN
+                IF ( State_Chm%Spc_Units == KG_SPECIES_PER_M2 ) THEN
                    REEVAPSO2 = GAINED * 96e+0_fp / 64e+0_fp &
                                * State_Grid%Area_M2(I,J)
                 ELSE
                    IF ( errPrint ) THEN
                       ErrorMsg= 'Unexpected species units: ' // &
-                                 TRIM( State_Chm%Spc_Units )
+                                 TRIM( UNIT_STR(State_Chm%Spc_Units) )
                       CALL GC_Error( ErrorMsg, RC, ThisLoc )
                    ENDIF
                    RETURN
@@ -4827,6 +4840,7 @@ CONTAINS
 #ifdef TOMAS
     USE TOMAS_MOD,      ONLY : IBINS, ICOMP, AQOXID
 #endif
+    USE UnitConv_Mod
 !
 ! !INPUT PARAMETERS:
 !
@@ -4956,13 +4970,13 @@ CONTAINS
              ! therefore converted to [kg] locally within AQOXID.
              ! WETLOSS is now [kg/m2] and so is multiplied
              ! by area prior to passing REEVAPSO2 to AQOXID (ewl, 9/30/15)
-             IF ( TRIM( State_Chm%Spc_Units ) .eq. 'kg/m2' ) THEN
+             IF ( State_Chm%Spc_Units == KG_SPECIES_PER_M2 ) THEN
                 REEVAPSO2 = - ( WETLOSS * 96e+0_fp / 64e+0_fp )              &
                             * State_Grid%Area_M2(I,J)
              ELSE
                 IF ( errPrint ) THEN
                    ErrorMsg = 'Unexpected species units: '                   &
-                               // TRIM(State_Chm%Spc_Units)
+                               // TRIM( UNIT_STR(State_Chm%Spc_Units) )
                    CALL GC_Error( ErrorMsg, RC, ThisLoc )
                 ENDIF
                 Spc => NULL()